+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/impl/BSMetaModelNameProvider.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -17,6 +17,7 @@
  */
 package com.intellij.idea.plugin.hybris.system.bean.meta.impl
 
+import com.intellij.idea.plugin.hybris.system.bean.meta.BSMetaHelper
 import com.intellij.idea.plugin.hybris.system.bean.model.*
 import com.intellij.idea.plugin.hybris.system.bean.model.Enum
 
@@ -24,7 +25,10 @@ object BSMetaModelNameProvider {
 
     fun extract(dom: Enum): String? = dom.clazz.stringValue?.takeIf { it.isNotBlank() }
     fun extract(dom: EnumValue): String? = dom.stringValue?.takeIf { it.isNotBlank() }
-    fun extract(dom: Bean): String? = dom.clazz.stringValue?.takeIf { it.isNotBlank() }
+    fun extract(dom: Bean): String? = dom.clazz.stringValue
+        ?.takeIf { it.isNotBlank() }
+        ?.let { BSMetaHelper.getBeanName(it) }
+
     fun extract(dom: Hint): String? = dom.name.stringValue?.takeIf { it.isNotBlank() }
     fun extract(dom: Property): String? = dom.name.stringValue?.takeIf { it.isNotBlank() }