+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/model/BSMetaBean.kt
@@ -1,6 +1,6 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
  * it under the terms of the GNU Lesser General Public License as
@@ -23,9 +23,13 @@ import com.intellij.idea.plugin.hybris.system.bean.model.BeanType
 interface BSMetaBean : BSMetaClassifier<Bean> {
     val type: BeanType
     val shortName: String?
+    val fullName: String?
+    val fullExtends: String?
+    val genericName: String?
     val description: String?
     val template: String?
     val extends: String?
+    val extendsGenericName: String?
     val deprecatedSince: String?
     val isDeprecated: Boolean
     val isAbstract: Boolean