+++ src/com/intellij/idea/plugin/hybris/system/bean/meta/model/impl/BSMetaBeanImpl.kt
@@ -1,10 +1,10 @@
 /*
  * This file is part of "SAP Commerce Developers Toolset" plugin for Intellij IDEA.
- * Copyright (C) 2019 EPAM Systems <hybrisideaplugin@epam.com>
+ * Copyright (C) 2019-2023 EPAM Systems <hybrisideaplugin@epam.com> and contributors
  *
  * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as 
- * published by the Free Software Foundation, either version 3 of the 
+ * it under the terms of the GNU Lesser General Public License as
+ * published by the Free Software Foundation, either version 3 of the
  * License, or (at your option) any later version.
  *
  * This program is distributed in the hope that it will be useful,
@@ -29,7 +29,7 @@ import com.intellij.util.xml.DomService
 internal class BSMetaBeanImpl(
     dom: Bean,
     override val module: Module,
-    override val name: String?,
+    override val name: String,
     override val isCustom: Boolean,
     override val imports: List<BSMetaImport>,
     override val annotations: List<BSMetaAnnotations>,
@@ -40,9 +40,14 @@ internal class BSMetaBeanImpl(
     override val domAnchor: DomAnchor<Bean> = DomService.getInstance().createAnchor(dom)
     override val description = dom.description.stringValue
     override val template = dom.template.stringValue
+    override val genericName = BSMetaHelper.getGenericName(dom.clazz.stringValue)
     override val extends = dom.extends.stringValue
+        ?.let { BSMetaHelper.getBeanName(it) }
+    override val extendsGenericName = BSMetaHelper.getGenericName(dom.extends.stringValue)
     override val type = dom.type.value ?: BeanType.BEAN
     override val shortName = BSMetaHelper.getShortName(name)
+    override val fullName = BSMetaHelper.getNameWithGeneric(name, genericName)
+    override val fullExtends = BSMetaHelper.getNameWithGeneric(extends, extendsGenericName)
     override val deprecatedSince = dom.deprecatedSince.stringValue
     override val isDeprecated = dom.deprecated.value
     override val isAbstract = dom.abstract.value
@@ -52,19 +57,22 @@ internal class BSMetaBeanImpl(
 
 }
 
-internal class BSGlobalMetaBeanImpl(localMeta: BSMetaBean)
-    : BSMetaSelfMerge<Bean, BSMetaBean>(localMeta), BSGlobalMetaBean {
+internal class BSGlobalMetaBeanImpl(localMeta: BSMetaBean) : BSMetaSelfMerge<Bean, BSMetaBean>(localMeta), BSGlobalMetaBean {
 
     override val hints = CaseInsensitive.CaseInsensitiveConcurrentHashMap<String, BSMetaHint>()
     override val properties = CaseInsensitive.CaseInsensitiveConcurrentHashMap<String, BSMetaProperty>()
     override val domAnchor = localMeta.domAnchor
     override val type = localMeta.type
     override val shortName = localMeta.shortName
+    override var genericName = localMeta.genericName
     override val module = localMeta.module
     override var template = localMeta.template
     override var description = localMeta.description
     override var deprecatedSince = localMeta.deprecatedSince
     override var extends = localMeta.extends
+    override var extendsGenericName = localMeta.extendsGenericName
+    override var fullName = localMeta.fullName
+    override var fullExtends = localMeta.fullExtends
     override val imports = ArrayList<BSMetaImport>()
     override val annotations = ArrayList<BSMetaAnnotations>()
     override var isDeprecated = localMeta.isDeprecated
@@ -72,10 +80,14 @@ internal class BSGlobalMetaBeanImpl(localMeta: BSMetaBean)
     override var isSuperEquals = localMeta.isSuperEquals
 
     override fun mergeInternally(localMeta: BSMetaBean) {
-        template?:let { template = localMeta.template }
-        extends?:let { extends = localMeta.extends }
-        description?:let { description = localMeta.description }
-        deprecatedSince?:let { deprecatedSince = localMeta.deprecatedSince }
+        template ?: let { template = localMeta.template }
+        extends ?: let { extends = localMeta.extends }
+        description ?: let { description = localMeta.description }
+        deprecatedSince ?: let { deprecatedSince = localMeta.deprecatedSince }
+        extendsGenericName ?: let { extendsGenericName = localMeta.extendsGenericName }
+        genericName ?: let { genericName = localMeta.genericName }
+        fullName ?: let { fullName = localMeta.fullName }
+        fullExtends ?: let { fullExtends = localMeta.fullExtends }
 
         if (localMeta.isDeprecated) isDeprecated = localMeta.isDeprecated
         if (localMeta.isAbstract) isAbstract = localMeta.isAbstract