+++ CHANGELOG.md
@@ -36,6 +36,7 @@
 ### Features
 - Added possibility to import Gradle KTS projects as modules [#534](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/534)
 - Improved folding for Relation tags in the `items.xml` [#555](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/555)
+- Improved Bean System support of the `extends` and `generics` [#570](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/570)
 
 ### Fixes
 - Inject `FlexibleSearch` into Kotlin String template [#535](https://github.com/epam/sap-commerce-intellij-idea-plugin/pull/535)